Subscribe once to all chat_message changes
